Add a script for simpler running of PolyTracker-instrumented `mutool`
This program began as an expansion of `mutool.sh`, but quickly grew
larger.

This program takes a PDF as input, runs a PolyTracker-instrumented
`mutool draw` on the PDF, and collects outputs & runtime metrics,
ultimately giving a JSON dictionary with the information.  For example:

    $ time ./mutool-draw --memory-limit 14g --time-limit 15m test.pdf
    {
        "input": {
            "filename": "test.pdf",
            "sha256sum": "2ad3d4c3b4f4370f7686f6ad8fa28f8ed35161f6e0df5aef880e878cefbbff2c",
            "size_bytes": 10162,
            "memory_limit": "14g",
            "time_limit": "15m"
        },
        "result": {
            "exit_code": 0,
            "memory_limited": false,
            "time_limited": false,
            "duration_seconds": 9.759987,
            "log_path": "test.pdf.log",
            "polytracker_output_path": "test.pdf.json"
        }
    }
    real    0m11.028s
    user    0m0.233s
    sys     0m0.086s

All of the `mutool` commands are done in a Docker container, which
should be automatically cleaned up afterward, unless the
`--skip-cleanup` option is given.