LabelDeq correction and performance improvement
There was a bug in LabelDeq that caused it to always choose the deque-path,
never the stack based path. It worked correctly, only slightly slower.
This commit fixes that and rewrites the impl to not use a variant, also
aiming to increase performance since this is one of the most common operations
called via TaintDAG::affects_control_flow.